Rename compnent parsing file
