+++ tests/types/test_date.py
@@ -6,7 +6,7 @@ from typing import Union
 import pytest
 from pydantic import ValidationError
 
-from ical._types import ComponentModel
+from ical.component import ComponentModel
 from ical.parsing.property import ParsedProperty
