+++ tests/types/test_date_time.py
@@ -6,7 +6,7 @@ from typing import Union
 import pytest
 from pydantic import ValidationError
 
-from ical._types import ComponentModel
+from ical.component import ComponentModel
 from ical.parsing.component import ParsedComponent
 from ical.parsing.property import ParsedProperty, ParsedPropertyParameter
 from ical.types.data_types import DATA_TYPE
