+++ tests/types/test_integer.py
@@ -3,7 +3,7 @@
 import pytest
 from pydantic import ValidationError
 
-from ical._types import ComponentModel
+from ical.component import ComponentModel
 from ical.parsing.property import ParsedProperty
