+++ tests/types/test_text.py
@@ -1,6 +1,6 @@
 """Tests for property values."""
 
-from ical._types import ComponentModel
+from ical.component import ComponentModel
 from ical.parsing.component import ParsedComponent
 from ical.parsing.property import ParsedProperty
